requirements.txt
ccxt==1.17.393
mysqlclient==1.3.13
matplotlib==3.0.0
bellmanford==0.2.1
numpy==1.15.2
networkx==2.2
pandas==0.23.4
tqdm==4.26.0
dill==0.2.7.1
pytest==3.8.2
pytest-cov==2.6.0
pytest-mock==1.10.0
neo4j-driver==1.6.2
ptvsd==4.1.4
asyncio
pytest-asyncio
neo4j-driver
asynctest